Check if config exists before copying the default

+++ docker/entrypoint.d/10-config.sh
@@ -1,2 +1,5 @@
 #!/bin/sh
-cp -au /app/rehome/resources/config/settings.yml "$SETTINGS_FILE_FOR_DYNACONF"
+if ! test -f "$SETTINGS_FILE_FOR_DYNACONF"; then
+    echo "Copying default settings.yml"
+    cp -au /app/rehome/resources/config/settings.yml "$SETTINGS_FILE_FOR_DYNACONF"
+fi
